+++ packages/alias/src/index.ts
@@ -1,3 +1,5 @@
+import path from 'path';
+
 import type { Plugin } from 'rollup';
 
 import type { ResolvedAlias, ResolverFunction, ResolverObject, RollupAliasOptions } from '../types';
@@ -97,7 +99,18 @@ export default function alias(options: RollupAliasOptions = {}): Plugin {
         updatedId,
         importer,
         Object.assign({ skipSelf: true }, resolveOptions)
-      ).then((resolved) => resolved || { id: updatedId });
+      ).then((resolved) => {
+        if (resolved) return resolved;
+
+        if (!path.isAbsolute(updatedId)) {
+          this.warn(
+            `rewrote ${importee} to ${updatedId} but was not an abolute path and was not handled by other plugins. ` +
+              `This will lead to duplicated modules for the same path. ` +
+              `To avoid duplicating modules, you should resolve to an absolute path.`
+          );
+        }
+        return { id: updatedId };
+      });
     }
   };
 }