+++ packages/alias/test/fixtures/warn.js
@@ -1,0 +1,2 @@
+import '@/warn-importee.js'
+import './folder/warn-importee.js'