+++ packages/esm-shim/test/fixtures/cjs-import-literal.js
@@ -1,0 +1,6 @@
+const dn = __dirname;
+
+module.exports = {
+  keyword: ' import',
+  dn
+};