+++ packages/replace/test/fixtures/form/ternary-operator/input.js
@@ -1,0 +1,3 @@
+/* eslint-disable */
+condition ? exprIfTrue : exprIfFalse;
+console.log(condition, exprIfTrue, exprIfFalse);