+++ packages/replace/test/fixtures/form/typescript-declare/_config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  description: "doesn't replace lvalue in typescript declare",
+  isTypescript: true,
+  options: {
+    'NAME': 'replaced',
+    preventAssignment: true
+  }
+};