+++ packages/run/types/index.d.ts
@@ -6,6 +6,7 @@ export interface RollupRunOptions extends ForkOptions {
   args?: readonly string[];
   options?: ForkOptions;
   allowRestarts?: boolean;
+  input?: string;
 }
 
 /**